Delete redundant file for Shapefile format parser.
