+++ engine/examples/src/main/scala/org/apache/spark/examples/sql/ShapeTypeTest.scala
@@ -43,23 +43,25 @@ object ShapeTypeTest {
       new StructField("shape", ShapeType, nullable = true)))
 
     // Polygon test
-    val shapefilePath = "/home/gefei/Downloads/gis/tl_2015_04_elsd/tl_2015_04_elsd.shp"
+    val shapefilePath = "/path/to/polygon-file"
     val gf = new GeometryFactory()
     val rdd = sc.parallelize(shapefile.Parser(shapefilePath)(gf)).map(r => (r.id, r.g))
     val df = rdd.toDF().show(5)    // toDF
 
-    val df2 = sqlContext.createDataFrame(rdd.map(org.apache.spark.sql.Row.fromTuple), schema).show(5)
-
+    val df2 = sqlContext.createDataFrame(rdd.map(org.apache.spark.sql.Row.fromTuple),
+      schema).show(5)
     rdd.toDF.registerTempTable("polygons")
-    sqlContext.sql("SELECT * FROM polygons").show(5) //sql test
+    sqlContext.sql("SELECT * FROM polygons").show(5) // sql test
 
 
     val rdd2 = sc.parallelize(1 to 10 map(i => Record(i, Point(Array(i, i, i)))))
     rdd2.toDF.show(5) // implicit convert
-    sqlContext.createDataFrame(rdd2.map(org.apache.spark.sql.Row.fromTuple), schema).show(5) //explicit covert
+    sqlContext.createDataFrame(rdd2.map(org.apache.spark.sql.Row.fromTuple),
+      schema).show(5) // explicit covert
     rdd2.toDF.registerTempTable("points")
-    sqlContext.sql("SELECT * FROM points").show(5) //sql
+    sqlContext.sql("SELECT * FROM points").show(5) // sql
 
     println("Finished.")
   }
 }
+
